requirements.txt
django==3.2.15
django-debug-toolbar==3.2.1
Pillow==8.2.0
environs[django]==9.3.2
django-phonenumber-field==7.0.2
phonenumbers==8.13.6
djangorestframework==3.14.0
requests==2.28.2
geopy==2.3.0
gunicorn==20.1.0
rollbar==0.16.3
psycopg2-binary==2.9.6

